Merged PR 5245: Updated viewer.properties file
Updated viewer.properties file as per -

https://github.com/mozilla/pdf.js/blob/master/l10n/en-US/viewer.properties

Included pdf.worker.js.

I'm not sure why this works, but it gets rid of all the warnings in the console when we generate a PDF. Also I have not been able to reproduce the blank pdf issue ever since I changed the file. But I'm not sure if it's down to this change.

Related work items: #305801