Merged PR 6084: Multiple SSO source POC
Trying out some stuff for being able to use different OAuth providers for HomeDesigner. Moving the AUTH_CONFIG injection out of common module is trivial (although not sure if it's even necessary).

With the observable, tried taking a route where I set the oauth provider in the route data, but ran into a chicken-and-egg situation because the route data is not available in a guard because it has not been activated yet. For now, using url + session storage, ~~but I'm thinking a more elegant approach might be to create an AuthSettings service within HomeDesigner, have that be the source for the observable stream, and then creating external and internal route guards that emit the settings in the canActivate method. This would allow us to configure which routes use which settings within the route config.~~ Edit: Added service and route guards to make it so oauthprovider can be set at the route level. Seems to be working pretty smoothly now.