Merged PR 6503: Sales Admin: Homesite Releases - An error occurred when deleting the home-sit...
Users could do an immediate release for a lot in another time zone, which would set the Release date to the next day. This causes the Edit/Delete buttons to show up for released lots under Home Site Releases.

Setting a time zone while saving an immediate release which would correctly set the date, and removing the offset while fetching it in the UI.

Related work items: #316491