Merged PR 6747: WI | 324205, 318724, 317993 | update out-of-sync spec pricing when selling home
When loading, checks if (a) user can sell stuff, (b) home is spec or model, (c) you're loading scenario or pending sales agreement, and (d) option prices have changed, and triggers a pricing update if necessary. Not sure whether it will try to trigger if you load a scenario on an already-sold lot, but the job pricing update has a check for that.

Also fixed the Cancel Spec and Spec Info issues while I was in there.

Might try and add a few unit tests later today if I have time.

Related work items: #317993, #318724, #324205