Merged PR 6795: Fixed subgroup not updating on filtered tree changes
Whenever the filtered tree is changes (ie deselecting previously contracted options), it is possible for the subgroup to change without the subgroupid changing. Not updating the selectedSubgroup will cause the child components to not update to use the new subgroup from the filtered tree, so this removes the check and will update these even if the subgroupId remains unchanged.

Related work items: #324713