Merged PR 6962: WI | 323808 | DT: Step 2 Cards Latency Issue
Cards in Step 2 were loading inconsistently due to a "value.replace is not a function" error. It appears the capitalCaseSpace pipe was taking in a number value based on the returned value from getSalesProgramType() in program-detail.component.ts. The code has been modified to return the string value from SalesProgramTypeEnum.

Related work items: #323808