Merged PR 7555: .NET Core fix - buyer disappearing after swap in pending agreement
There was a `slice(1)` being done to the `buyersAfterSwap` object which seems to have been causing the swapped buyer to drop off the list. After removing the slice, all buyers remain in the grid when swapping, and from what I've tested it doesn't look like this is breaking anything else. But there may be a test case I missed, do we know why that slice was there to begin with?