Merged PR 7905: Remove logic causing incorrect prices for options
There was some arithmetic with list prices that was causing negative values and also the "last" choice to take the combined sum of all of its prior choice prices in the choice path. With the recent option mapping changes, this code now seems to be unnecessary.

The original intent may have been to restore the correct list price in the event the price changed at some point between COs, but this code did not appear to be addressing that exactly.