Merged PR 8058: when option is removed, remove attributes as well
This is the only way I could think of to do this. But there is a potential gap, I believe if one of the attribute groups is removed from the option, it's not going to remove that attribute group from the choice since it's using the option data to determine what to remove.

It *might* be possible to put something in an effect that looks at JobPlanOption records to determine what is actually associated with the removed option... May need to explore that for a possible "phase 2" of this fix.

Related work items: #335588