Merged PR 8339: Merged PR 8337: Bug | 337390 | load financial communities when opening scenario
Merged PR 8337: Bug | 337390 | load financial communities when opening scenario

I'd like to do away with the salesCommunity property within ScenarioState, it seems a little redundant. Then we'd be able to get rid of the includeFinancialCommunities parameter altogether. Unfortunately the scope of that change is too much for the timeframe we're working with...

Related work items: #337390

Cherry-picked from commit `84cdb29e`.

Related work items: #337390