Merged PR 8825: WI | 331700 | DT: Unable to Print Floor Plan
#Overview
In situations where users bypass the Floorplan tab, or an error occurred, the floor plan attachment SVG images are not saved to the database, resulting in a blank floor plan report. By adding a hidden floor plan component onto the summary screen, a save operation is triggered and the user can see the floor plan report properly.

#Changes
- Moved floor-plan.component.ts to shared module.
- Updated scenario-summary.component.ts to include hidden floor-plan component.
- Added a flag in floor-plan component to force save without needing the agreement to be editable.
- Added emitter to update the summaryImages for the report data, so the user doesn't have to refresh the browser to see the report correctly.

Related work items: #331700