Merged PR 9376: WI | 346204 | CA: Division Choices tab is blank
In QA, the Division Choices tab is showing the error `Object(...) is not a function` on div-choices-panel.component.ts:135, in reference to the forkJoin. I noticed that the forkJoin import doesn't follow the same path as the rest of our components, so maybe that difference has something to do with the error.

Locally, the Division Choices tab has been working fine before and after the change, so I'm not sure if it's a configuration issue somewhere specific to QA. Open to ideas if there are any suggestions, or we can see if changing the import fixes it first.

Related work items: #346204