Merged PR 9842: WI | 347028 | DT: Unable to Save Change Order due to Null in DecisionPointLbl
This fixes the error and sets the proper labels, but I'm not sure if using the choiceLabel is the best way to go about this... the divChoiceCatalogId was different between the original and the updated choice, so the choiceLabel was the only thing that remained consistent. Is there a better way to retrieve that updated choice?

Related work items: #347028