Merged PR 10639: WI | 351642 | DT: IFP not showing 2nd Floor
Due to some timing of callbacks and lack of data, static SVGs were not appearing properly on agreements that could not be edited. The floor plan component has been updated to better handle the `canEditAgreement` flag and now has a separate callback handler for static images in the event that the IFP is not initialized, or existing images are not present.

Related work items: #351642