Merged PR 10792: WI | 358009 | DT: Set timeOfSaleOptionPrices to empty array if null on the choice
During testing in QA, an issue was found when selecting an attribute on a construction change order for a newly-selected choice. This action triggered a save operation for TimeOfSaleOptionPrice records, but it was null. This resulted in being mapped to an array containing an undefined object, which caused the error. The code has been updated to default the property to an empty array if not found.

Related work items: #358009