Merged PR 11515: Create tag on prod deployment approval
Build task to automatically create a git tag when a production deployment is complete. Yesterday I noticed it's sometimes difficult to figure out exactly which source version is in production if the production build is older than the retention period (I think 3 months). This creates a tag whose name is the build version (e.g., Phd-SalesAdmin-UI_1.0.0.11441). It also deletes old tags for that build version (currently have it set up to keep the 3 most recent tags for the build definition).

Sample:
![image.png](https://pultevsts.visualstudio.com/6d26cbbb-94e5-4db9-ac14-039e14cd6ff9/_apis/git/repositories/599adcf9-26a5-4eb9-9264-f574911edf59/pullRequests/11515/attachments/image.png)