Merged PR 11529: WI | 358693 | SA: Handle IsWebSaleable EDH field in Sales Programs
Adding support for the new IsWebSaleable column in EDH Sales.SalesProgram, which will replace the ThoBuyerClosingCostId and ThoDiscountFlatAmountId columns in PHD com.FinancialCommunityInfo. The code is currently storing data in both tables since existing PHD data needs to be migrated. References to the PHD table are currently marked for deletion, but not removed.

Related work items: #358693