Merged PR 11896: Disable THO Validation build step
Got rid of most of the errors by changing how the moment and lodash imports are done. However, there is still some tslib related stuff that fails on building THO, and I suspect it could be because the output from an Angular 13 library build is incompatible with an Angular 10 application. Tried quickly upgrading THO to Angular 13, but ran into issues building the sass files.

May revisit this after THO is upgraded to Angular 13, or may try to put together something where it can use the old Angular 11 version, but that will be quite complicated.