Merged PR 12754: WI | 367177 | Fix plans not loading when market level feature switch does not exist
Until a plan is selected we don't have the financialCommunityId. The feature is enabled using that id - 8362. Re-instated the original logic that defaults isPhdLiteEnabled to true when the financialCommunityId is unknown. The feature switch check will happen again once the plan is selected and everything will work as expected at that point.

Related work items: #367177