Merged PR 14678: DT: Fix for pricing discrepancy
This fixes an issue with incorrect pricing on some choices related to removed options. I think that there was overreaching logic for determining if a choice had a removed option, which might have been to account for missing time-of-sale records. That issue, due to the Cancel button incorrectly deleting some time-of-sale data, was resolved, so it appears that the original "hasRemovedOption" logic is working as expected now.

Related work items: #376671