Merged PR 15170: WI | 378729 | DT: Fix for incorrect price after cancelling CO
In a CO, when Choice A is originally selected, but the selection is changed to Choice B, new options/pricing is brought onto Choice A. When the CO is cancelled, the new pricing remains on Choice A, and the only way to restore original pricing is to create another CO and do nothing but cancel it. It seems that the lockedInChoice was never restored to Choice A when cancelling the first CO, so a flag has been added for that specific action to maintain the original price.

Related work items: #378729