Merged PR 15914: WI | #380895 | Update People component to set trustName
WI | #380895 | Update People component to set trustName in the store so that trust will be saved to state.SalesAgreement. Previously the TrustName would save to the changeOrder and not the SalesAgreement. This caused the People status bar indicator to not fill at the top of the page.

Related work items: #380895