Merged PR 17507: WI | 387129 | DT - Generate Agreement button scrolls through house images
This is a prod issue where a user can't generate agreement on tablet when the green button is clicked. The root cause is that the carousel control overlaps the green button in smaller screen. Setting the width and height for the carousel arrows seems to fix the issue.

Related work items: #387129