Merged PR 18875: WI | 392019 | Issue: Choice is not disabled in same session "cancel" of Change Order
When dealing with multiple tree versions and choices/options changing between them, it seems the only way to restore those options in the same session, after cancelling a CO, is to force a reload of the SAG so it gets the correct tree data. Not sure if that's advisable, let me know if there could be an alternative solution.

Related work items: #384993, #392019