Merged PR 18855: Bug | 388716 | update choice list price calculation method
This ditches the strategy of looking at every possible combination of choices that could directly impact pricing, and instead looks at all the options where the choice is the maxSortOrderChoice, and tries to satisfy each option mapping combination in order of highest/lowest price.

Couldn't think of a good algorithm to come up with ordered combinations, if anybody remembers any from their university math classes let me know; I just went with iterating over a large range of numbers and returning combinations as we find them. Might not be ideal, but it seems to only take 40 seconds to do the calculation for the "problem" plan, so it might be good enough.

Related work items: #388716